client: Use tip of master in repo insted of released pip package
In Client, do not use the released version of labgrid, but build from the
tip of master.
The dockerfile might stop working because the versions are not pinned
in the setup file, but this is something to solved separately.
In client switch container inheritance to use Python container.
Get rid of virtualenv, as this is not needed in a docker env.

Signed-off-by: Kasper Revsbech <krev@triax.com>